Bug #26691: Only use file path for files
The documentation, though poor, implies that there are three options for
invoking the Add-AzIoTDeviceProvisioningServiceCertificate cmdlet.

One is with the path to a .pem formatted X509 certificate file.

One is with a .cer (raw) file.

The last is the problem.  It implies that if neither .pem nor .cer
file paths are supplied, then the -Path will be used as a base64 value.

Looking at the code, that also appears to be the intent.

I adjusted the code to have a separate private variable which stores the
user resolved path derived from the Path passed in.

In the third case, use the original path versus the resolved path.