Feature #26664: Add IsVerified option to Add-AzIoTDeviceProvisioningServiceCertificate
Currently, the command does not support the IsVerified option which
makes the certificate validation unnecessary.

KLUDGE WARNING:  THIS CODE IS WRONG!!!  I FORCE THE API VERSION TO A
NEWER VERSION WITHOUT USING THE AUTOREST MAGIC TO REGENERATE WHATEVER IS
BEING GENERATED.  I COULD NOT FIND THE DOCUMENTATION AND REALLY DIDN'T
CARE.

I verified that with this change, the API call now has the isVerified
property set in the body of the PUT request and the certificate is added
with its verified flag set.