Bug #26669: Adjust Set-IoTDeviceProvisioningServiceCertificate to name parameters, pass certificate raw bytes.
The Set-IoTDeviceProvisioningServiceCertificate command currently wrongly
passes the certificate provided to it as an option as the stringified
version (either immediate, or by reading the path) which, because of
positional parameters being used, gets passed as the certificatename
parameter.  It really should be provided as the certificateRawBytes
parameter.

I also adjust the call to use named parameters for everything to reduce
chances of this problem in the future.